images: Add demo-image-docker
NVIDIA docker containers after Jetpack 5 expect libraries to be provided
inside the container instead of passed through to the host.  So there's
a benefit to introduce a smaller image specifically for docker support
which doesn't include all the content in demo-image-full

Note that even with this change some of the larger containers aren't
possible to load from NGC on EMMC parts or parts with 16GB or less
root filesystem.

Signed-off-by: Dan Walkes <danwalkes@trellis-logic.com>