Add symlink for engine file
For some reason not understood deepstream 7.0 creates
an engine file at a different path than it is referenced
in the config file.

Add a hack symlink between the two file names so the engine
created by gstnvinfer is found by the pipeline on startup.

This greatly decreases the restart time of the process
since the engine doesn't need to get rebuilt every run.

Signed-off-by: Dan Walkes <danwalkes@trellis-logic.com>