swupdate: add L4TLauncher update to swupdate
Based on issues at [1] and [2] and discussions in the monthly
meetings at [3], [4], and [5] add an update to the L4TLauncher code
contained in the ESP partition boot directory.

Since this will not be brick-proof, make it conditional on
update-if-different logic also used for bootloader upgrades.  We
may want even stricter logic on when to use this copy if you are
especially concerned about brick-proof upgrades.

I considered using NVIDIA's scheme to use esp_alt partition and
the procedure outlined at [6].  However I decided not to go this
route because:
1) I think it still is not brick-proof, since the swapping of names
and partition GUIDs at [7] will not be atomic.
2) It adds complication to the variable update process, since the
ESP capsule needs to be aded to the partition to trigger capsule update,
so the esp alt partition swap would need to happen early in the update
process and then the alt partition would need modification before boot.

1: https://matrix.to/#/!YBfWVpJwNVtkmqVCPS:gitter.im/$WHyIkCEONvEDLukleV7gKtlUDp2-uMIHEflK7nRxw3Q?via=gitter.im&via=matrix.org&via=3dvisionlabs.com
2: https://github.com/orgs/OE4T/discussions/1874
3: https://github.com/OE4T/meta-tegra/wiki/OE4T-Meeting-Notes-2024%E2%80%9003%E2%80%9014
4: https://github.com/OE4T/meta-tegra/wiki/OE4T-Meeting-Notes-2024%E2%80%9005%E2%80%9009
5: https://github.com/OE4T/meta-tegra/wiki/OE4T-Meeting-Notes-2025%E2%80%9004%E2%80%9010
6: https://docs.nvidia.com/jetson/archives/r36.4.3/DeveloperGuide/SD/SoftwarePackagesAndTheUpdateMechanism.html#updating-the-esp-partition
7: https://github.com/Trellis-Logic/ota-tools-jetson/blob/f2418f3cfcadb8b9f3a0de166a40177256ae106a/Linux_for_Tegra/tools/ota_tools/version_upgrade/nv_ota_update_alt_part.func#L389-L398

Signed-off-by: Dan Walkes <danwalkes@trellis-logic.com>